+++ parallel-second-implementation/em_algorithm.h
@@ -4,9 +4,9 @@
 #include "constants.h"
 
 // em algorithm
-void em_train(Sample *samples, float *mean, float *cov, float *weights, float *p_val, int process_samples, int process_rank);
+void em_train(Sample *samples, double *mean, double *cov, double *weights, double *p_val, int process_samples, int process_rank);
 
 // initialize mean, cov and weights 
-void initialize(float *mean, float *cov, float *weights);
+void initialize(double *mean, double *cov, double *weights);
 
 #endif
