+++ parallel-second-implementation/m_step.h
@@ -4,6 +4,6 @@
 #include "constants.h"
 
 // run maximization step
-void m_step_parallel(float *local_p_val, Sample *samples, float *mean, float *cov, float *weights, int process_rank, int process_samples);
+void m_step_parallel(double *local_p_val, Sample *samples, double *mean, double *cov, double *weights, int process_rank, int process_samples);
 
 #endif
